pyproject.toml
[project]
name = "watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "kc-nremc", email = "k_courtney@nremc.com" }
]
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
]

[project.optional-dependencies]
cachelib = [
    "flask-session>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
]

[tool.isort]
profile = "black"
